service/tradfri/requirements.txt
twisted
cyclone
docopt
python-dateutil
pytradfri[async]==6.4.0

rdflib==4.2.2
rdflib-jsonld==0.4.0

cycloneerr
patchablegraph==0.11.0
rdfdb==0.21.0
standardservice==0.6.0

